+++ smt/utils/kriging_utils.py
@@ -334,7 +334,7 @@ def gower_componentwise_distances(X, y=None, xtypes=None):
 
     n_samples, n_features = X_num.shape
     n_nonzero_cross_dist = n_samples * (n_samples - 1) // 2
-    ij = np.zeros((n_nonzero_cross_dist, 2), dtype=np.int)
+    ij = np.zeros((n_nonzero_cross_dist, 2), dtype=np.int32)
     D_num = np.zeros((n_nonzero_cross_dist, n_features))
     ll_1 = 0
     if y is None:
@@ -373,7 +373,7 @@ def gower_componentwise_distances(X, y=None, xtypes=None):
         D[:, np.logical_not(cat_features)] = D_num
         D[:, cat_features] = D_cat
 
-        return D, ij.astype(np.int), X_cont
+        return D, ij.astype(np.int32), X_cont
     else:
         D = X_norma[:, np.newaxis, :] - Y_norma[np.newaxis, :, :]
         D = D.reshape((-1, X.shape[1]))
